+++ .github/release-notes.md
@@ -1,2 +1,3 @@
 * Show self-inflicted deaths in killfeed
 * Fix error when steam path is incorrect
+* Store past player avatars